refactor(direct-control): extract ready read types
Refs: openspec/changes/civ7-support-direct-control-modularization

Moves only ready-unit input/operation/nearby/promotion/result types into packages/civ7-direct-control/src/play/ready/unit.ts, unit-move-preview input/result types into packages/civ7-direct-control/src/play/ready/move-preview.ts, and ready-city input/operation/production/town-focus/population-placement/result types into packages/civ7-direct-control/src/play/ready/city.ts. Keeps public facade type re-exports in index.ts.

Preserves public type contracts and does not change ready-unit, unit-move-preview, or ready-city source strings, validator/default/bounds behavior, read-only/no-send behavior, relationship-label policy, production/town-focus/population-placement semantics, runtime proof status, telemetry, AI ingestion, hotseat runtime proof, CLI semantic projection, Effect/oRPC procedure-core work, or Task 2.9.4 matrix-row acceptance. Diplomacy/narrative closeout, operation, public procedure schema, and telemetry type ownership remain pending separate owner slices.

Validation:
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test -- ready-unit-view.test.ts unit-move-preview.test.ts ready-city-view.test.ts public-api.test.ts
- bun run --cwd packages/cli test -- game/play/ready-unit.test.ts game/play/unit-move-preview.test.ts game/play/ready-city.test.ts game/play/priorities.test.ts
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- bun run --cwd apps/mapgen-studio check
- git diff --check
- ready type ownership scan and ready module no-cycle import scan

Local package/type relocation proof only; no runtime/live-game, ready read behavior, relationship classification, CLI semantic projection, telemetry, AI-ingestion, hotseat product-path, or procedure-core proof claimed.
