+++ openspec/changes/civ7-support-direct-control-modularization/tasks.md
@@ -456,6 +456,13 @@ runtime/direct-control claims.
         procedure schemas, operation/proof telemetry, hotseat runtime proof, AI
         ingestion, CLI semantic projection, Effect/oRPC procedure-core work,
         and Task 2.9.4 matrix-row acceptance pending.
+  - [x] 4.8.22 Extract ready read type owners while keeping public facade type
+        re-exports in `index.ts`, preserving ready-unit, unit-move-preview, and
+        ready-city public contracts and conservative relationship-label policy
+        while leaving diplomacy/narrative closeout, operation, public procedure
+        schemas, operation/proof telemetry, hotseat runtime proof, AI ingestion,
+        CLI semantic projection, Effect/oRPC procedure-core work, and Task
+        2.9.4 matrix-row acceptance pending.
 - [ ] 4.9 Import or explicitly cite oRPC architecture authority before any
       oRPC implementation; current support branch has no tracked
       `.agents/skills/civ7-orpc-control-architecture` or
